+++ src/Eav/Infra/Doctrine/ObjectMappings.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace MsgPhp\Eav\Infra\Doctrine;
 
 use MsgPhp\Domain\Infra\Doctrine\{MappingConfig, ObjectMappingProviderInterface};
-use MsgPhp\Eav\Entity\{AttributeValue, Features};
+use MsgPhp\Eav\Entity\{Attribute, AttributeValue, Features, Fields};
 
 /**
  * @author Roland Franssen <franssen.roland@gmail.com>
@@ -23,6 +23,15 @@ final class ObjectMappings implements ObjectMappingProviderInterface
                 'cascade' => ['all'],
                 'joinColumns' => [
                     ['nullable' => false, 'onDelete' => 'CASCADE'],
+                ],
+            ],
+        ];
+        yield Fields\AttributeField::class => [
+            'attribute' => [
+                'type' => self::TYPE_MANY_TO_ONE,
+                'targetEntity' => Attribute::class,
+                'joinColumns' => [
+                    ['nullable' => false],
                 ],
             ],
         ];
