+++ src/Eav/Tests/Infra/Doctrine/ObjectMappingsTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace MsgPhp\Eav\Tests\Infra\Doctrine;
 
 use MsgPhp\Domain\Infra\Doctrine\MappingConfig;
-use MsgPhp\Eav\Entity\Fields\AttributesField;
 use MsgPhp\Eav\Infra\Doctrine\ObjectMappings;
 use PHPUnit\Framework\TestCase;
 
@@ -16,8 +15,6 @@ final class ObjectMappingsTest extends TestCase
         $available = array_flip(array_map(function (string $file): string {
             return 'MsgPhp\\Eav\\Entity\\'.basename(\dirname($file)).'\\'.basename($file, '.php');
         }, array_merge(glob(\dirname(__DIR__, 3).'/Entity/Features/*.php'), glob(\dirname(__DIR__, 3).'/Entity/Fields/*.php'))));
-        unset($available[AttributesField::class]);
-
         $mappings = ObjectMappings::provideObjectMappings(new MappingConfig([]));
         $mappings = array_keys($mappings instanceof \Traversable ? iterator_to_array($mappings) : $mappings);
         sort($mappings);
