+++ src/UserEav/Entity/Fields/AttributeValuesField.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace MsgPhp\User\Entity\Fields;
 
 use MsgPhp\Domain\{DomainCollection, DomainCollectionInterface};
-use MsgPhp\Eav\Entity\Fields\AttributesField;
 use MsgPhp\User\Entity\UserAttributeValue;
 
 /**
@@ -13,9 +12,9 @@ use MsgPhp\User\Entity\UserAttributeValue;
  */
 trait AttributeValuesField
 {
-    use AttributesField;
-
-    /** @var iterable|UserAttributeValue[] */
+    /**
+     * @var iterable|UserAttributeValue[]
+     */
     private $attributeValues = [];
 
     /**
